#1339 - Modularization of Action Bar
- Started reorganizing the action bar
- Merged CurationPanel into CurationPage to make the CurationPage more similar to the other pages derived from AnnotationPageBase faciliating the action bar refactoring
