+++ webanno-api-annotation/src/main/java/de/tudarmstadt/ukp/clarin/webanno/api/annotation/actionbar/DocumentNavigator.html
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<!--
+#Copyright 2019
+#Ubiquitous Knowledge Processing (UKP) Lab and FG Language Technology
+#Technische Universität Darmstadt
+#
+#Licensed under the Apache License, Version 2.0 (the "License");
+#you may not use this file except in compliance with the License.
+#You may obtain a copy of the License at
+# 
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+#Unless required by applicable law or agreed to in writing, software
+#distributed under the License is distributed on an "AS IS" BASIS,
+#WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#See the License for the specific language governing permissions and
+#limitations under the License.
+-->
+<html xmlns:wicket="http://wicket.apache.org">
+<body>
+  <wicket:panel>
+    <div class="btn-group">
+      <button wicket:id="showPreviousDocument" class="btn btn-default" type="button">
+        <i class="fa fa-step-backward" aria-hidden="true"></i>
+      </button>
+      <button wicket:id="showNextDocument" class="btn btn-default" type="button">
+        <i class="fa fa-step-forward" aria-hidden="true"></i>
+      </button>
+    </div>
+  </wicket:panel>
+</body>
+</html>
