+++ webanno-api-annotation/src/main/java/de/tudarmstadt/ukp/clarin/webanno/api/annotation/page/AnnotationPageBase.properties
@@ -16,10 +16,6 @@
 toggleScriptDirection.label=LTR/RTL
 toggleScriptDirection.icon=images/text_width.png
 
-showPreviousDocument.label=Prev.
-showPreviousDocument.tooltip=Press Shift+PgUp
-showPreviousDocument.icon=images/document_back.png
-
 showNextDocument.label=Next
 showNextDocument.tooltip=Press Shift+PgDown
 showNextDocument.icon=images/document_next.png
