+++ webanno-ui-annotation/src/main/java/de/tudarmstadt/ukp/clarin/webanno/ui/annotation/AnnotationPage.html
@@ -41,15 +41,10 @@
                       src="images/drawer_open.png" title="Open" /><br />Open
                   </a></li>
                   
-                  <li wicket:id="showPreviousDocument" />
-                  <li wicket:id="showNextDocument" />
                   <li wicket:id="showExportDialog" />
                   <li wicket:id="showPreferencesDialog" />
                 </ul>
               </fieldset>
-            </li>
-  
-            <li wicket:id="pageNavigator">
             </li>
   
             <li>
@@ -83,21 +78,24 @@
             </li>
           </ul>
         </div>
-        <div class="locationbar">
-          <span wicket:id="documentNamePanel"></span> <span style="float: right;">
-          <span wicket:id="numberOfPages"></span>
-          </span>
-        </div>
       </div>
       
       <div class="annotation-editor-container flex-content flex-h-container flex-gutter flex-no-initial-gutter">
         <div wicket:id="leftSidebar" class="flex-sidebar flex-v-container flex-gutter"></div>
-        <div class="flex-content flex-v-container flex-gutter" wicket:enclosure="editor">
+        <div wicket:id="centerArea" class="flex-content flex-v-container flex-gutter">
           <div class="flex-content panel panel-primary panel-flex">
             <div class="panel-heading">
-              <h3 class="panel-title">Annotation</h3>
+              <h3 class="panel-title">
+                <span wicket:id="documentNamePanel"/> 
+                <span wicket:id="numberOfPages" style="float: right;"/>
+              </h3>
             </div>
-            <div wicket:id="editor" class="panel-body flex-v-container" style="padding: 0px;">
+            <div class="panel-body flex-v-container" style="padding: 0px;">
+              <nav class="flex-h-container">
+                <span wicket:id="documentNavigator" />
+                <span wicket:id="pageNavigator" />
+              </nav>
+              <div wicket:id="editor" class="flex-content flex-v-container" style="padding: 0px;"/>
             </div>
           </div>
         </div>
