+++ webanno-ui-core/src/main/java/de/tudarmstadt/ukp/clarin/webanno/ui/core/css/theme/custom.less
@@ -304,3 +304,15 @@ a.disabled {
 .bootstrap-select.bs-container .dropdown-menu {
   z-index: 21060 !important;
 }
+
+.annotation-editor-container {
+	nav {
+		padding: 5px;
+		background-color: grey;
+		
+		> * {
+			padding-left: 15px;
+			padding-right: 15px;
+		}
+	}	
+}
