--- webanno-ui-curation/src/main/java/de/tudarmstadt/ukp/clarin/webanno/ui/curation/component/CurationPanel.html
@@ -1,153 +1,0 @@
-<!DOCTYPE html>
-<!--
-#Copyright 2012
-#Ubiquitous Knowledge Processing (UKP) Lab and FG Language Technology
-#Technische Universität Darmstadt
-#
-#Licensed under the Apache License, Version 2.0 (the "License");
-#you may not use this file except in compliance with the License.
-#You may obtain a copy of the License at
-# 
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-#Unless required by applicable law or agreed to in writing, software
-#distributed under the License is distributed on an "AS IS" BASIS,
-#WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#See the License for the specific language governing permissions and
-#limitations under the License.
--->
-<html xmlns:wicket="http://wicket.apache.org">
-<head>
-<wicket:head>
-  <style type="text/css">
-    .curation-sentence-item {
-      	margin: 2px 2px;
-      	padding: 2px 2px;
-    }
-    
-    .curation-sentence-number {
-      border: 1px solid white;
-      border-radius: 4px 4px 4px 4px;
-      -moz-border-radius: 4px 4px 4px 4px;
-      -webkit-border-radius: 4px 4px 4px 4px;
-      	margin: 1px 1px;
-      padding: 1px 1px;
-      display: inline-block;
-      	min-width: 4em;
-      	max-width: 7.9em;
-      text-align: center;
-    }
-
-    .curation-sentence-number a {
-      display: block;
-      text-decoration: none !important;
-      color: black !important;
-      border-radius: 4px 4px 4px 4px;
-      -moz-border-radius: 4px 4px 4px 4px;
-      -webkit-border-radius: 4px 4px 4px 4px;
-    }
-    
-    .curation-sentence-list {
-      -webkit-box-sizing: border-box;
-      -moz-box-sizing: border-box;
-      	box-sizing: border-box;
-      	width: 100%;
-      	height: 550px;
-      	overflow: scroll;
-    }
-
-    .hidden-view {
-      	background-color: #DFDFDF;
-    }
-    
-    .in-range {
-      	border: 1px solid black;
-    }
-    
-    .out-range {
-      	border: 1px solid white;
-    }
-    
-    .agree {
-    	
-    }
-    
-    .disagree {
-      	background-color: red;
-    }
-    
-    .current {
-      box-shadow: 0px 0px 5px 1px rgba(231,166,26,1);
-        border: 1px solid rgba(231,166,26,1);
-    }
-    
-    .suggestion-panel-body {
-      overflow-x: auto;
-    }
-  </style>
-</wicket:head>
-</head>
-<body>
-  <wicket:panel>
-    <div class="annotation-editor-container flex-content flex-h-container flex-gutter">
-
-      <!-- 
-         We cannot put a 'wicket:enclosure="sentencesListView"' on this div because this 
-         inevitably causes the sentence list to "jump up" when re-rendering it (i.e. when paging
-         or when selecting a sentence).
-       -->
-      <div wicket:id="sentenceListContainer" class="flex-sidebar flex-v-container flex-gutter">
-        <div class="flex-content panel panel-default panel-flex">
-          <div class="panel-heading">
-            <h3 class="panel-title">Sentences</h3>
-          </div>
-          <div class="scrolling panel-body">
-            <div wicket:id="sentencesListView" style="border-collapse: collapse;">
-              <span wicket:id="sentencesList" class="curation-sentence-number">
-                <a wicket:id="sentenceNumber"></a>
-              </span>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      <div class="flex-content flex-v-container">
-        <div class="flex-content flex-v-container flex-gutter">
-          <div class="flex-content flex-v-container flex-gutter flex-only-internal-gutter" style="flex: 0.75;">
-            <div class="flex-content panel panel-primary panel-flex" wicket:enclosure="mergeView">
-              <div class="panel-heading">
-                <h3 class="panel-title">Annotation</h3>
-              </div>
-              <div wicket:id="mergeView" class="panel-body text-center flex-v-container"></div>
-            </div>
-    
-            <div class="panel panel-default" wicket:enclosure="crossSentAnnoView">
-              <div class="panel-heading">
-                <h3 class="panel-title">Invisible annotations on other pages</h3>
-              </div>
-              <div class="panel-body">
-                <span wicket:id="crossSentAnnoView">
-                  <span wicket:id="crossSentAnnoList">
-                    <span wicket:id="crossAnnoSent"></span>
-                  </span>
-                </span>
-              </div>
-            </div>
-          </div>
-          
-          <div wicket:id="suggestionViewPanel" class="scrolling flex-content">
-            <div wicket:id="contextMenu" style="width: 250px;"></div>
-            <div class="sentence-list-view" wicket:id="sentenceListView">
-              <div wicket:id="sentence"></div>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      <div wicket:id="rightSidebar" class="flex-sidebar flex-v-container">
-        <wicket:container wicket:id="annotationDetailEditorPanel"/>
-      </div>
-    </div>
-  </wicket:panel>
-</body>
-</html>
